Split _model.py into cohesive modules
`_model.py` had grown into a 1800-line module spanning nine unrelated
responsibilities. Relocate each group into a dedicated module (pure
moves, no logic changes):

- _typing.py: shared type aliases and the numeric tolerance
- _linalg.py: gaussian_basis and the matrix/covariance kernels
- _penalties.py: l1/group penalties and their proximal operators
- _forward.py: ForwardModel
- _data.py: RegressionData and covariate_from_stim
- _initialization.py: find_mu, wls, mne_initialization
- _solver.py: FitHistory, _evaluate_objective, Solver, find_mu_range

_model.py now holds only NCRF, NCRFModel, and NCRFResult. The public
package surface is unchanged; __init__ and _ncrf re-export from the new
locations, and the test/doc references to moved names are updated.
